Changed CategoryRepository, ItemCategoryRepository, OrderRepository and ItemRepository interfaces to use Long as ID.

+++ src/main/java/com/answerdigital/benhession/academy/answerkingweek2/repositories/ItemCategoryRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ItemCategoryRepository extends CrudRepository<ItemCategory, Integer> {
-}
+public interface ItemCategoryRepository extends CrudRepository<ItemCategory, Long>
+{ }
